doppel-core/pyproject.toml
[project]
name = "doppel-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chatterbox-tts",
    "onnx==1.18",
    "torchcodec>=0.8.1",
]

[tool.uv.extra-build-dependencies]
pkuseg = ["numpy>=1.24.0,<1.26.0"]

[tool.uv.workspace]
members = [
    "chatterbox",
]

[tool.uv.sources]
chatterbox-tts = { workspace = true }
